\name{COUSCOus-internal}
\title{Internal COUSCOus functions}
\alias{generate.rho}
\alias{precision}
\alias{prediction}
\alias{preprocessing}
\alias{shrink.S}
\description{Internal COUSCOus functions}
\author{
Reda Rawi \cr
Maintainers: Reda Rawi <redarawi@gmx.de>
}
\details{These are not intended for usage. Please refer to \code{COUSCOus()}
}
